Rename: Move styles folder to router folder inside
