+++ app/msg-box/[roomId]/styles/modalCase.module.scss
