+++ app/msg-box/[roomId]/styles/modalRules.module.scss
