+++ app/msg-box/[roomId]/styles/page.module.scss
