+++ app/msg-box/[roomId]/styles/ruleListItem.module.scss
