+++ app/msg-box/layout.tsx
@@ -1,4 +1,4 @@
-import style from './layout.module.scss'
+import style from './styles/layout.module.scss'
 export default function layout({ children }: { children: React.ReactNode }) {
   return <section className={style.container}>{children}</section>
 }
