+++ app/msg-box/styles/layout.module.scss
