+++ app/msg-box/styles/message.module.scss
