+++ app/msg-box/styles/messagePreview.module.scss
